Split up and clean up backend sample scene
