+++ tests/pr/backend/backend.d3d12.cc
@@ -1,6 +1,8 @@
 #ifdef PR_BACKEND_D3D12
 #include <nexus/test.hh>
 
+#include <iostream>
+
 #include "backend_common.hh"
 
 #include <phantasm-renderer/backend/d3d12/BackendD3D12.hh>
