+++ tests/pr/backend/backend_common.cc
@@ -1,137 +1,26 @@
 #include "backend_common.hh"
 
+#include <iostream>
+
+#include <clean-core/capped_vector.hh>
+#include <clean-core/defer.hh>
+
+#include <typed-geometry/tg.hh>
+
+#include <arcana-incubator/asset-loading/image_loader.hh>
+#include <arcana-incubator/asset-loading/mesh_loader.hh>
+
+#include <phantasm-renderer/backend/assets/vertex_attrib_info.hh>
+#include <phantasm-renderer/backend/command_stream.hh>
 #include <phantasm-renderer/backend/detail/byte_util.hh>
 #include <phantasm-renderer/backend/detail/format_size.hh>
 #include <phantasm-renderer/backend/detail/unique_buffer.hh>
+#include <phantasm-renderer/backend/device_tentative/timer.hh>
+#include <phantasm-renderer/backend/device_tentative/window.hh>
+#include <phantasm-renderer/default_config.hh>
 
-void pr_test::copy_mipmaps_to_texture(pr::backend::command_stream_writer& writer,
-                                      pr::backend::handle::resource upload_buffer,
-                                      std::byte* upload_buffer_map,
-                                      pr::backend::handle::resource dest_texture,
-                                      pr::backend::format format,
-                                      const inc::assets::image_size& img_size,
-                                      inc::assets::image_data& img_data,
-                                      bool use_d3d12_per_row_alingment,
-                                      bool limit_to_first_mip)
-{
-    using namespace pr::backend;
-    auto const bytes_per_pixel = detail::pr_format_size_bytes(format);
-
-    CC_RUNTIME_ASSERT(img_size.array_size == 1 && "array upload unimplemented");
-    cmd::copy_buffer_to_texture command;
-    command.source = upload_buffer;
-    command.destination = dest_texture;
-    command.texture_format = format;
-    command.dest_mip_size = limit_to_first_mip ? 1 : img_size.num_mipmaps;
-
-    auto accumulated_offset_bytes = 0u;
-
-    for (auto a = 0u; a < img_size.array_size; ++a)
-    {
-        command.dest_array_index = a;
-
-        // copy all the mip slices into the offsets specified by the footprint structure
-        //
-        for (auto mip = 0u; mip < command.dest_mip_size; ++mip)
-        {
-            auto const mip_width = cc::max(unsigned(tg::floor(img_size.width / tg::pow(2.f, float(mip)))), 1u);
-            auto const mip_height = cc::max(unsigned(tg::floor(img_size.height / tg::pow(2.f, float(mip)))), 1u);
-
-            command.dest_width = mip_width;
-            command.dest_height = mip_height;
-            command.source_offset = accumulated_offset_bytes;
-            command.dest_mip_index = mip;
-
-            writer.add_command(command);
-
-            auto const mip_row_size_bytes = bytes_per_pixel * command.dest_width;
-            auto mip_row_stride_bytes = mip_row_size_bytes;
-
-            if (use_d3d12_per_row_alingment)
-                // MIP maps are 256-byte aligned per row in d3d12
-                mip_row_stride_bytes = mem::align_up(mip_row_stride_bytes, 256);
-
-            auto const mip_offset_bytes = mip_row_stride_bytes * command.dest_height;
-            accumulated_offset_bytes += mip_offset_bytes;
-
-            inc::assets::rowwise_copy(img_data, upload_buffer_map + command.source_offset, mip_row_stride_bytes, mip_row_size_bytes, command.dest_height);
-
-            // if applicable, write the next mip level
-            if (mip + 1 < command.dest_mip_size)
-                inc::assets::write_mipmap(img_data, command.dest_width, command.dest_height);
-        }
-    }
-}
-
-unsigned pr_test::get_mipmap_upload_size(pr::backend::format format, const inc::assets::image_size& img_size)
-{
-    using namespace pr::backend;
-    auto const bytes_per_pixel = detail::pr_format_size_bytes(format);
-    auto res_bytes = 0u;
-
-    for (auto a = 0u; a < img_size.array_size; ++a)
-    {
-        for (auto mip = 0u; mip < img_size.num_mipmaps; ++mip)
-        {
-            auto const mip_width = cc::max(unsigned(tg::floor(img_size.width / tg::pow(2.f, float(mip)))), 1u);
-            auto const mip_height = cc::max(unsigned(tg::floor(img_size.height / tg::pow(2.f, float(mip)))), 1u);
-
-            auto const row_pitch = mem::align_up(bytes_per_pixel * mip_width, 256);
-            auto const custom_offset = row_pitch * mip_height;
-            res_bytes += custom_offset;
-        }
-    }
-
-    //    uint32_t num_rows[D3D12_REQ_MIP_LEVELS] = {0};
-    //    UINT64 row_sizes_in_bytes[D3D12_REQ_MIP_LEVELS] = {0};
-    //    D3D12_PLACED_SUBRESOURCE_FOOTPRINT placed_subres_tex2d[D3D12_REQ_MIP_LEVELS];
-    //    size_t res;
-    //    {
-    //        auto const res_desc = CD3DX12_RESOURCE_DESC::Tex2D(util::to_dxgi_format(format), img_size.width, img_size.height, 1, UINT16(img_size.num_mipmaps));
-    //        device.GetCopyableFootprints(&res_desc, 0, img_size.num_mipmaps, 0, placed_subres_tex2d, num_rows, row_sizes_in_bytes, &res);
-    //    }
-
-    //    std::cout << "Custom: " << res_bytes << ", D3D12: " << res << std::endl;
-
-    return res_bytes;
-}
-
-
-pr::backend::format pr_test::get_texture_format(bool hdr, unsigned num_channels)
-{
-    using pf = pr::backend::format;
-    if (hdr)
-    {
-        switch (num_channels)
-        {
-        case 4:
-        default:
-            return pf::rgba16f;
-        case 3:
-            return pf::rgb16f;
-        case 2:
-            return pf::rg16f;
-        case 1:
-            return pf::r16f;
-        }
-    }
-    else
-    {
-        switch (num_channels)
-        {
-        case 4:
-        default:
-            return pf::rgba8un;
-        case 3:
-            return pf::rgb8u;
-        case 2:
-            return pf::rg8u;
-        case 1:
-            return pf::r8u;
-        }
-    }
-}
-
+#include "sample_scene.hh"
+#include "texture_util.hh"
 
 void pr_test::run_sample(pr::backend::Backend& backend, const pr::backend::backend_config& config, sample_config const& sample_config)
 {
@@ -223,7 +112,7 @@ void pr_test::run_sample(pr::backend::Backend& backend, const pr::backend::backe
             resources.mat_metallic = load_texture(pr_test::sample_metallic_path);
             resources.mat_roughness = load_texture(pr_test::sample_roughness_path);
 
-            resources.ibl_lut = load_texture("res/pr/liveness_sample/texture/brdf_lut.png", 2, true, true); // backend.createTexture(format::rg16f, 256, 256, 1);
+            resources.ibl_lut = load_texture("res/pr/liveness_sample/texture/brdf_lut.png", 2, true, true);
             resources.ibl_specular = backend.createTexture(format::rgba16f, 256, 256, 0, texture_dimension::t2d, 6);
             resources.ibl_irradiance = backend.createTexture(format::rgba16f, 256, 256, 1, texture_dimension::t2d, 6);
